backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
Werkzeug==2.3.0
fpdf2==2.7.0
gunicorn==21.2.0
